selftests: mlxsw: qos_lib: Split devlink_..._set() into save & set [xxx]
Changing pool type from static to dynamic causes reinterpretation of
threshold values. They therefore need to be saved before pool type is
changed, then the pool type can be changed, and then the new values need
to be set up.

Therefore set cannot subsume save, because it would be saving the wrong
thing, with possibly a nonsensical value, and restore would then fail to
restore the nonsensical value.

Thus extract a _save() from each of the relevant _set()'s so that it is
possible to put the pool mode change between them.

Signed-off-by: Petr Machata <petrm@mellanox.com>