mlxsw: reg: Extract a helper for trap registration
On Spectrum-1, timestamps for PTP packets are delivered through queues
of ingress and egress timestamps. There are two event traps
corresponding to activity on each of those queues. This mechanism is
absent on Spectrum-2, and therefore the traps should only be registered
on Spectrum-1.

Extract out of mlxsw_sp_traps_init() a generic helper,
__mlxsw_sp_traps_init(), and likewise with _fini(). The new helpers will
later be called with Spectrum-1-specific traps.

Signed-off-by: Petr Machata <petrm@mellanox.com>