mlxsw: spectrum: Add support for Spectrum-1-specific traps
On Spectrum-1, timestamps for PTP packets are delivered through queues
of ingress and egress timestamps. There are two event traps
corresponding to activity on each of those queues. This mechanism is
absent on Spectrum-2, and therefore the traps should only be registered
on Spectrum-1.

Add a new listener array for Spectrum-1 traps. Register the trap
listeners in mlxsw_sp1_init() and unregister in mlxsw_sp1_fini() (which
add).

The listener array is empty for now, the events will be added in a later
patch.

Signed-off-by: Petr Machata <petrm@mellanox.com>