mlxsw: Add a new module to define tracepoints
Currently, tracepoints are defined in the ATCAM module, because it was
the only tracepoint user in mlxsw. A follow-up commit will add
tracepoints to the PTP subsystem, and it is likely more will arrive in
the future. Defining tracepoints in the ATCAM object, while not wrong,
does not seem to be appropriate anymore. Move the definitions to a
dedicated object file.

Signed-off-by: Petr Machata <petrm@mellanox.com>