mlxsw: spectrum_ptp: Set the PTP shaper parameters
Set the PTP shaper parameters during the ptp_init(). For different speeds,
there are different parameters.

When the port speed changes and PTP shaper is enabled, the firmware changes
the ETS shaper values according to the PTP shaper parameters for this new
speed.

The PTP shaper parameters array is left empty for now, will be filled in a
follow-up patch.

Signed-off-by: Shalom Toledo <shalomt@mellanox.com>
Reviewed-by: Petr Machata <petrm@mellanox.com>